Make init load custom props
This code has been modified in a way that it loads props from oem.prop.
The user has to create a file named oem.prop in the
root directory (i.e. '/') with the content being the props from the
below paths:
    - /ocm/system-properties/config.prop
    - /ocm/system-properties/cust.prop

keeping in mind not to have these set of props:
    - ro.build.description
    - ro.build.version.incremental
    - ro.build.tags
    - ro.build.fingerprint
    - ro.vendor.build.fingerprint
    - ro.bootimage.build.fingerprint

in the file oem.prop

In this way, user can load his/her region and carrier specific props for
stuffs such as VoLTE, VoWiFi, etc; without overriding the un-necessary
props mentioned above.

Signed-off-by: Shashank Verma <shashank.verma2002@gmail.com>