yoshino: Add NetworkSwitcher
This app is responsible for switching network mode between 3G and LTE.

After IMS registration and subsequent reboots, the modem goes into restart loops. This can be fixed by connecting to 3G first and then moving up to LTE upon boot completion. Why that is the case remains a mystery for now. This app takes care of this issue automatically, switching to 3G upon shutdown and LTE after boot completed.

Signed-off-by: Shashank Verma <shashank.verma2002@gmail.com>