NetworkSwitcher: Implementation reboot prompt
When IMS is setup by selecting the required modem mbn, it is necessary to boot the phone in 3G and then switching back to LTE upon boot completion - toggling to 3G and back to LTE at boot is not enough. This appears to be required for proper IMS registration as the modem will enter a reboot loop otherwise.  So when the following two conditions are met, reboot the phone so that the NetworkSwitcher can take care of this by switching to 3G on shutdown:

1) a non-default mbn has been used by modem-switcher
   AND
2) there are no preferences if selected network stored for the NetworkSwitcher (indicating eithet a first boot or clearing if its data)

Signed-off-by: Shashank Verma <shashank.verma2002@gmail.com>