OpenCustomizationSelector: Tighten the SimID matching
As is, the CustomizationSelector will try to find a best match when no exact match is found.

This can lead to some unwanted side-effects. For example, if an MCC but no MNC matches, the first entry with matching MCC will be chosen. This can lead to a wrong CustID selection and hence the wrong modem modification as well.

Tighten up the SimID matching logic, by requiring _at least_ an MCC + MCC match. Also, only check for a ServiceProvider, IMSI and Gid1 match if both MCC and MNC match.

Finally, remove the Gid2 logic completely, since Gid2 is not supported in AOSP and LOS.

Signed-off-by: Shashank Verma <shashank.verma2002@gmail.com>